requirements.txt
python-dotenv
python-telegram-bot
opencv-python
ultralytics
pyqt5